refactor lsp client
