+++ Composer/packages/lib/code-editor/package.json
@@ -53,7 +53,7 @@
     "ts-loader": "^5.3.3",
     "url-loader": "^1.1.2",
     "webpack": "^4.30.0",
-    "webpack-cli": "^3.3.2",
+    "webpack-cli": "^3.3.10",
     "webpack-dev-server": "^3.8.2"
   },
   "dependencies": {
@@ -61,15 +61,17 @@
     "@bfc/lg-languageserver": "*",
     "@bfcomposer/monaco-editor-webpack-plugin": "^1.7.2",
     "@bfcomposer/react-monaco-editor": "0.30.12",
-    "@types/node": "^12.12.5",
+    "@types/node": "^12.12.8",
+    "@types/ws": "^6.0.3",
+    "lodash": "^4.17.15",
     "lodash.throttle": "^4.1.1",
     "monaco-editor-core": "0.17.0",
-    "monaco-languageclient": "^0.10.2",
+    "monaco-languageclient": "^0.10.0",
     "net": "^1.0.2",
     "normalize-url": "^4.5.0",
-    "reconnecting-websocket": "^4.2.0",
+    "reconnecting-websocket": "^3.2.2",
     "vscode": "^1.1.36",
     "vscode-ws-jsonrpc": "^0.1.1",
-    "lodash": "^4.17.15"
+    "ws": "^5.0.0"
   }
 }
