+++ Composer/packages/lib/code-editor/src/LgEditor.tsx
@@ -2,8 +2,17 @@
 // Licensed under the MIT License.
 
 import React from 'react';
-import { startSampleClient, registerLGLanguage } from '@bfc/language-client';
+import { registerLGLanguage } from '@bfc/language-client';
+import { listen, MessageConnection } from 'vscode-ws-jsonrpc';
+import {
+  MonacoLanguageClient,
+  CloseAction,
+  ErrorAction,
+  MonacoServices,
+  createConnection,
+} from 'monaco-languageclient';
 
+import { createUrl, createWebSocket } from './utils/lspUtil';
 import { RichEditor, RichEditorProps } from './RichEditor';
 
 const LG_HELP =
@@ -36,6 +45,62 @@ const defaultLGServer = {
   path: '/lgServer',
 };
 
+async function initializeDocuments(languageClient, lgOption) {
+  if (lgOption && lgOption.inline) {
+    await languageClient.onReady();
+    languageClient.sendRequest('initializeDocuments', lgOption);
+  }
+}
+
+function createLanguageClient(connection: MessageConnection): MonacoLanguageClient {
+  return new MonacoLanguageClient({
+    name: 'LG Language Client',
+    clientOptions: {
+      // use a language id as a document selector
+      documentSelector: ['botbuilderlg'],
+      // disable the default error handler
+      errorHandler: {
+        error: () => ErrorAction.Continue,
+        closed: () => CloseAction.DoNotRestart,
+      },
+    },
+    // create a language client connection from the JSON RPC connection on demand
+    connectionProvider: {
+      get: (errorHandler, closeHandler) => {
+        return Promise.resolve(createConnection(connection, errorHandler, closeHandler));
+      },
+    },
+  });
+}
+
+let monacoServiceInstance;
+
+function startLSPClient(editor, lgServer, lgOption) {
+  // install Monaco language client services
+  if (!monacoServiceInstance) {
+    monacoServiceInstance = MonacoServices.install(editor);
+  }
+
+  lgOption.uri = editor.getModel().uri._formatted;
+
+  // create the web socket
+  const url = createUrl(lgServer);
+  const webSocket = createWebSocket(url);
+  // listen when the web socket is opened
+  listen({
+    webSocket,
+    onConnection: connection => {
+      // create and start the language client
+      const languageClient = createLanguageClient(connection);
+      // send full content
+      initializeDocuments(languageClient, lgOption);
+      const disposable = languageClient.start();
+      connection.onClose(() => disposable.dispose());
+    },
+  });
+  return editor;
+}
+
 export function LgEditor(props: LGLSPEditorProps) {
   const options = {
     quickSuggestions: true,
@@ -51,7 +116,7 @@ export function LgEditor(props: LGLSPEditorProps) {
     }
   };
   const editorDidMount = (editor, monaco) => {
-    startSampleClient(editor, lgServer, lgOption);
+    startLSPClient(editor, lgServer, lgOption);
     if (typeof props.editorDidMount === 'function') {
       return props.editorDidMount(editor, monaco);
     }
