+++ Composer/packages/tools/language-client/demo/src/util.ts
@@ -1,0 +1,27 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+const normalizeUrl = require('normalize-url');
+const ReconnectingWebSocket = require('reconnecting-websocket');
+
+export function createUrl(server: { [key: string]: string } | string): string {
+  if (typeof server === 'string') {
+    return normalizeUrl(server).replace(/^http/, 'ws');
+  }
+  const { host, hostname = location.hostname, port = location.port, path = '/' } = server;
+  const protocol = location.protocol === 'https:' ? 'wss' : 'ws';
+  const endHost = host || `${hostname}:${port}`;
+  return normalizeUrl(`${protocol}://${endHost}/${path}`);
+}
+
+export function createWebSocket(url: string): WebSocket {
+  const socketOptions = {
+    maxReconnectionDelay: 10000,
+    minReconnectionDelay: 1000,
+    reconnectionDelayGrowFactor: 1.3,
+    connectionTimeout: 10000,
+    maxRetries: Infinity,
+    debug: false,
+  };
+  return new ReconnectingWebSocket(url, [], socketOptions);
+}
