+++ Composer/packages/tools/language-client/src/index.ts
@@ -1,5 +1,4 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-export * from './startSampleClient';
 export * from './languages';
